refactor: ProductionData using page factory pattern
chore: cleanup
