+++ bciers/apps/reporting/src/tests/components/products/ProductionDataPage.test.tsx
@@ -1,6 +1,7 @@
 import { getProductionData } from "@bciers/actions/api";
-import ProductionData from "@reporting/src/app/components/products/ProductionData";
+import ProductionDataPage from "@reporting/src/app/components/products/ProductionDataPage";
 import { render, screen } from "@testing-library/react";
+import { HasFacilityId } from "@reporting/src/app/utils/defaultPageFactoryTypes";
 
 vi.mock("@bciers/actions/api", () => ({
   getProductionData: vi.fn(),
@@ -10,6 +11,12 @@ vi.mock("@reporting/src/app/utils/getOrderedActivities", () => ({
 }));
 
 const getProductionDataMock = getProductionData as ReturnType<typeof vi.fn>;
+
+// 🏷 Constants
+const props: HasFacilityId = {
+  version_id: 1,
+  facility_id: "abc",
+};
 
 describe("The Production Data component", () => {
   beforeEach(() => {
@@ -22,7 +29,7 @@ describe("The Production Data component", () => {
       report_products: [],
     });
 
-    render(await ProductionData({ facility_id: "abc", report_version_id: 1 }));
+    render(await ProductionDataPage(props));
     expect(getProductionData).toHaveBeenCalledWith(1, "abc");
   });
 
@@ -34,7 +41,7 @@ describe("The Production Data component", () => {
       ],
       report_products: [],
     });
-    render(await ProductionData({ facility_id: "abc", report_version_id: 1 }));
+    render(await ProductionDataPage(props));
 
     expect(screen.getAllByText(/production data/i)).toHaveLength(2); // One for the page, one for the tasklist
     expect(
@@ -55,7 +62,7 @@ describe("The Production Data component", () => {
       ],
     });
 
-    render(await ProductionData({ facility_id: "abc", report_version_id: 1 }));
+    render(await ProductionDataPage(props));
     expect(screen.getByText("testProduct")).toBeInTheDocument();
     expect(screen.getByText("Unit")).toBeInTheDocument();
     expect(screen.getByText("tonnes of tests")).toBeInTheDocument();
